script: Add support to generate loader's table and extension headers for VulkanSC.
Preprocessor 'VulkanSC' is added for the vulkansc builds.
Using this preprocessor to deactivate the functions that are not supported on VulkanSC

Loader: Added the preprocessor check for VULKANSC in the loader code.
Most of the change is simple preprocessor guard word addition that only
compiles the relevant code paths for VULKANSC.
Loader emulation is not required since device_groups and
get_physical_device_features2 are core in VulkanSC

BUILD.md : Updated the VulkanSC build commands

generate the vksc files with:

  cmake --build . --target VulkanLoader_generated_source