Fix a regkey-related test failure
Some code in windows_shim.cpp tries to look at the seventh character in
a regkey name in order to determine which of a small set of lists to
iterate through.

This breaks when changing the name of the VulkanDriverName regkey to
VulkanSCDriverName.

To fix this, use wcsstr to locate a more detailed part of the regkey
name in order to identify it.